+++ wpilibjExamples/src/main/java/edu/wpi/first/wpilibj/examples/examples.json
@@ -229,7 +229,7 @@
     "commandversion": 2
   },
   {
-    "name": "Motor Control With Encoder",
+    "name": "Motor Control",
     "description": "Demonstrate controlling a single motor with a Joystick and displaying the net movement of the motor using an encoder.",
     "tags": [
       "Robot and Motor",
@@ -239,7 +239,7 @@
       "Joystick",
       "Complete List"
     ],
-    "foldername": "motorcontrolencoder",
+    "foldername": "motorcontrol",
     "gradlebase": "java",
     "mainclass": "Main",
     "commandversion": 2
